v1.12.8
Improvements:

Updated the GopenPGP cryptographic library to version 2 for better security and improved stability
Added a warning that announces to users that we will drop compatibility with iOS10
Improved the experience when adding badly formatted email addresses as a recipient of an email
Improved error handling when sending emails to badly formatted email addresses
Improved the process of creating, updating, saving and recovering drafts
Improved the synchronization of messages
improved various aspects of UI/UX for better clarity

Fixes:

Sometimes user would receive duplicated push notifications with the second one showing a generic message saying "You received a new message!"
Sometimes a message sending error would be displayed saying "Message sending failed. Please try again: EOF"
Sometimes a message sending error would be displayed saying "Message sending failed. The package is required."
In some cases, the received message would appear encrypted only in the iOS app
In some cases, an "invalid access token" message would be displayed after opening the app
When starting with 0, the Two-Factor Authentication code would not be accepted
The back button on the create PIN flow did not work
Fixed various crashes that could happened when sending a message
Fixed a crash that occurred when deleting emails from the trash folder
Fixed a crash that sometimes occurred when the app is in the background
Fixed a crash that sometimes occurred when selecting a phone number from the message body
Fixed a crash that sometimes occurred when opening the app from a push notification
Fixed a bug whereby the refresh spinner would keep spinning forever
Fixed various issues related to displaying the updated list of emails
Fixed a bug whereby text contained in angled brackets would be badly displayed
Fixed a bug in the report bug flow that prevented users from properly submitting the report
Fixed a bug whereby a queue management issue leading to the spinner spinning forever
Fixed various bugs whereby the download of the message list would be delayed or incomplete