Fix missing native swipe-back gesture - from left edge to right
Hiding the native back bar button item in SwiftUI disables native
swipe-back gesture which is a default gesture in iOS
(https://stackoverflow.com/a/58045361).

To keep the gesture it's necessary to extend `UINavigationController` with
`UIGestureRecognizerDelegate` protocol and configure
interactivePopGestureRecognizer's delegate which is used under
the hood by the SwiftUI for navigation. The other option would be to
use `UINavigationController` wrapped up with `UIHostingController` but
this requires more work.